+++ source/Core/Controller/BaseController.php
@@ -8,6 +8,8 @@ namespace OxidEsales\EshopCommunity\Core\Controller;
 
 use OxidEsales\EshopCommunity\Core\ShopVersion;
 use Psr\Container\ContainerInterface;
+use OxidEsales\EshopCommunity\Internal\ShopEvents\AfterRequestProcessedEvent;
+
 
 /**
  * Base view class. Collects and passes data to template engine, sets some global
@@ -532,6 +534,7 @@ class BaseController extends \OxidEsales\Eshop\Core\Base
             if (method_exists($this, $sFunction)) {
                 $sNewAction = $this->$sFunction();
                 self::$_blExecuted = true;
+                $this->dispatchEvent(new AfterRequestProcessedEvent);
 
                 if (isset($sNewAction)) {
                     $this->_executeNewAction($sNewAction);
@@ -601,6 +604,8 @@ class BaseController extends \OxidEsales\Eshop\Core\Base
             }
 
             $this->onExecuteNewAction();
+
+            $this->dispatchEvent(new AfterRequestProcessedEvent);
 
             //#M341 do not add redirect parameter
             \OxidEsales\Eshop\Core\Registry::getUtils()->redirect($url, (bool) $myConfig->getRequestParameter('redirected'), 302);